Fix NextJS routing issues with `redirect()` in a server action
